requirements.txt
argparse
os
io
numpy
shutil
pillow
unrealcv